Move Headphone UiState class
